feat(misc): updating downloadedfromhies and medicaldatasource
deleting duplicated files in api and refactoring to point to core

Refs: #1350
