+++ packages/api/src/command/medical/document/document-reconvert.ts
@@ -13,8 +13,8 @@ import { Patient } from "../../../domain/medical/patient";
 import { convertCDAToFHIR } from "../../../external/fhir-converter/converter";
 import { getDocuments as getDocumentsFromFHIRServer } from "../../../external/fhir/document/get-documents";
 import { countResources } from "../../../external/fhir/patient/count-resources";
-import { downloadedFromHIEs } from "../../../external/fhir/shared";
-import { getMetriportContent } from "../../../external/fhir/shared/extensions/metriport";
+import { downloadedFromHIEs } from "@metriport/core/external/fhir/shared/index";
+import { getMetriportContent } from "@metriport/core/external/fhir/shared/extensions/metriport";
 import { Config } from "../../../shared/config";
 import { errorToString } from "../../../shared/log";
 import { formatNumber } from "@metriport/shared/common/numbers";
