+++ packages/api/src/command/medical/document/document-redownload.ts
@@ -21,8 +21,8 @@ import {
 import { hasCommonwellContent, isCommonwellContent } from "../../../external/commonwell/extension";
 import { makeFhirApi } from "../../../external/fhir/api/api-factory";
 import { getPatientId } from "../../../external/fhir/patient";
-import { downloadedFromHIEs } from "../../../external/fhir/shared";
-import { isMetriportContent } from "../../../external/fhir/shared/extensions/metriport";
+import { downloadedFromHIEs } from "@metriport/core/external/fhir/shared/index";
+import { isMetriportContent } from "@metriport/core/external/fhir/shared/extensions/metriport";
 import { getAllPages } from "../../../external/fhir/shared/paginated";
 import { PatientModel } from "../../../models/medical/patient";
 import { filterTruthy } from "../../../shared/filter-map-utils";
