+++ packages/api/src/external/commonwell/extension.ts
@@ -1,7 +1,7 @@
 import { DocumentReference, DocumentReferenceContent, Extension } from "@medplum/fhirtypes";
 import { MedicalDataSource } from "@metriport/api-sdk";
-import { dataSourceExtensionDefaults } from "../fhir/shared/extensions/extension";
-import { MetriportDataSourceExtension } from "../fhir/shared/extensions/metriport";
+import { dataSourceExtensionDefaults } from "@metriport/core/external/fhir/shared/extensions/extension";
+import { MetriportDataSourceExtension } from "@metriport/core/external/fhir/shared/extensions/metriport";
 
 export const cwExtension: MetriportDataSourceExtension = {
   ...dataSourceExtensionDefaults,
