--- packages/api/src/external/fhir/shared/extensions/base-extension.ts
@@ -1,1 +1,0 @@
-export const BASE_EXTENSION_URL = "https://public.metriport.com/fhir/StructureDefinition";
