+++ packages/api/src/routes/helpers/default-error-handler.ts
@@ -3,7 +3,7 @@ import { ErrorRequestHandler } from "express";
 import httpStatus from "http-status";
 import { ZodError } from "zod";
 import MetriportError from "../../errors/metriport-error";
-import { getDetailFromOutcomeError } from "../../external/fhir/shared";
+import { getDetailFromOutcomeError } from "@metriport/core/external/fhir/shared/index";
 import { isClientError } from "../../shared/http";
 import { capture } from "../../shared/notifications";
 import { httpResponseBody } from "../util";
