+++ packages/core/src/external/index.ts
@@ -1,5 +1,6 @@
 export enum MedicalDataSource {
   COMMONWELL = "COMMONWELL",
+  CAREQUALITY = "CAREQUALITY",
 }
 
 export function isMedicalDataSource(s?: string | null): s is MedicalDataSource {
@@ -12,4 +13,5 @@ export const HL7OID = "2.16.840.1.113883";
 
 export const MedicalDataSourceOid: Record<MedicalDataSource, string> = {
   [MedicalDataSource.COMMONWELL]: `${HL7OID}.3.3330`,
+  [MedicalDataSource.CAREQUALITY]: `${HL7OID}.3.3330`,
 };
